REF: Move `diagnostics
